feat: release 0.4.1 with connect command, validator fixtures, and doc updates
- Add connect/disconnect command (JS + Python) for claude-desktop, cursor, vscode, windsurf
- Add USEPASO_AUTH_TOKEN env var support via .env file loading (JS + Python)
- Fix lint-staged hook: use local node_modules binaries for eslint and prettier
- Fix ANSI color detection: use stdout.isTTY instead of stderr.isTTY !== false
- Fix JSON output parity: add compact separators to Python inspect/validate commands
- Add test-fixtures for validator edge cases (ssrf, consent, string-length)
- Add test-fixtures for MCP server (consent-required, constraints)
- Bump version to 0.4.1 in JS and Python packages
- Update spec: document string-length validation rules and warnings section
- Update all READMEs: fix Cursor config, add connect/disconnect, update CLI table

+++ packages/paso-py/usepaso/types.py
@@ -76,7 +76,7 @@ class PasoOutput:
 class PasoConstraint:
     max_per_hour: Optional[int] = None
     max_per_request: Optional[int] = None
-    max_value: Optional[int] = None
+    max_value: Optional[float] = None
     allowed_values: Optional[List[Any]] = None
     requires_field: Optional[str] = None
     description: Optional[str] = None
